upload avatar to account

+++ src/modules/auth/services/update_avatar_account_service.ts
@@ -13,24 +13,29 @@ type AvatarAccountModel = {
 export class UpdateAvatarAccountService {
   async uploadImage(data: AvatarAccountModel): Promise<void> {
     const accountRepository = AppDataSource.getRepository(AccountModel)
-    const account: any = accountRepository.findOneBy({ id: data.id })
+    const account: any = await accountRepository.findOneBy({ id: data.id })
+    console.log('email: ' + account.email)
+    console.log('file: ' + data.avatar)
 
     if (!account) throw new AppError('Account not found')
 
-    if (account.avatar) {
+    if (account.avatar && account.avatar === 'avatar.default') {
       const avatarAccount = path.join(uploadConfig.directory, account.avatar)
       const avatarAccountExists = await fs.promises.stat(avatarAccount)
 
       if (avatarAccountExists) {
+        console.log('avatar', account.avatar)
         await fs.promises.unlink(avatarAccount) // apagar o arquivo
       }
 
+      console.log('arquivo ', data.avatar)
       await accountRepository
         .createQueryBuilder()
         .update(AccountModel)
         .set({
-          avatar: data.id
-        }).where({ id: data.id }).execute()
+          avatar: data.avatar
+          // updatedAt: new Date()
+        }).where({ id: account.id }).execute()
     }
   }
 }
